refactor(examples): splitting src of CustomFilters example; split
